feat: update SDK library and bytecode compilers
The CMakeLists.txt file in the SDK source directory was updated to include a new path for header files. In addition, the `Compile` method of `_NotImplementedYet` class in `bytecode_compiler.cpp` was modified to take a `ScriptRecipe&`. The same modification was made for the `Compile` method of both x86_64_compiler.cpp and x86_compiler.cpp. All methods now return an empty vector.

This change will allow developers to compile their scripts using these compilers with ease.